requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
cssselect2==0.7.0
dnspython==2.6.1
docx2pdf==0.1.8
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
install==1.3.5
Jinja2==3.1.4
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==8.4.0
numpy==1.26.4
openpyxl==3.1.3
orjson==3.10.3
oscrypto==1.3.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyExcelerate==0.11.0
Pygments==2.18.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
svglib==1.5.1
tinycss2==1.3.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
xhtml2pdf==0.2.15